+++ docs/scripts/webpack.config.babel.js
@@ -14,6 +14,7 @@ export default {
           `?${ JSON.stringify({
             context: path.join(config.cwd, config.source),
             basepath: config.baseUrl.pathname,
+            description: {},
             feedsOptions: {
               title: pkg.name,
               site_url: pkg.homepage,