+++ src/md-collection-loader/index.js
@@ -1,17 +1,13 @@
 /*
 Example
-
 ```md
   ---
   title: Test
   key: value
   ---
-
   _Md_ content
 ```
-
 return a object:
-
 ```json
 {
   __filename: ...
@@ -26,7 +22,6 @@ return a object:
   rawBody: "_Md_ content",
   raw: {initial content},
 }
-
 ```
  */
 import path from "path"
@@ -39,6 +34,7 @@ import toUri from "../_utils/to-uri"
 import enhanceCollection from "../enhance-collection"
 import feed from "./feed"
 import cache from "./cache"
+import description from "./description"
 
 // also in builder/server.js
 const fileExtensionRE = /\.(.*)+$/
@@ -89,7 +85,7 @@ module.exports = function(input) {
     __resourceUrl: joinUri(basepath, resourceUrl),
     __dataUrl: joinUri(basepath, dataUrl),
   }
-  const mdObject = {
+  let mdObject = {
     head: {
       ...defaultHead,
       ...parsed.data,
@@ -99,6 +95,8 @@ module.exports = function(input) {
     raw: parsed.orig,
     ...metadata,
   }
+
+  mdObject = description(mdObject, query.description)
 
   if (!this.emitFile) {
     throw new Error("emitFile is required from module system")